Update creation of date attribute to be in Toolbar
